+++ src/dirichlet-multinomial-distribution.cpp
@@ -38,7 +38,7 @@ NumericVector cpp_ddirmnom(
   
   int Nmax = std::max({
     static_cast<long int>(x.nrow()),
-    size.length(),
+    static_cast<long int>(size.length()),
     static_cast<long int>(alpha.nrow())
   });
